formbuilder replace with formio

+++ app/Http/Controllers/Admin/Applications/ApplicationPipelineController.php
@@ -6,7 +6,9 @@ use App\DataTables\Admin\Applications\ApplicationPipelinesDataTable;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Applications\Pipeline\PipelineStoreRequest;
 use App\Models\ApplicationPipeline;
+use App\Models\ApplicationPipelineStage;
 use Google\Service\TrafficDirectorService\Pipe;
+use Illuminate\Support\Facades\DB;
 
 class ApplicationPipelineController extends Controller
 {
@@ -25,7 +27,13 @@ class ApplicationPipelineController extends Controller
 
   public function store(PipelineStoreRequest $request)
   {
-    ApplicationPipeline::create($request->validated());
+    $pipeline = ApplicationPipeline::create($request->validated());
+
+    $is_default_set = false;
+
+    foreach ($request->stages as $i => $stage) {
+      $pipeline->stages()->create(['is_default' => !$is_default_set && boolVal(@$stage['is_default'])] + $stage);
+    }
 
     return $this->sendRes('Pipeline created successfully', ['event' => 'table_reload', 'table_id' => 'application-pipelines-table', 'close' => 'globalModal']);
   }
@@ -39,13 +47,32 @@ class ApplicationPipelineController extends Controller
 
   public function update(PipelineStoreRequest $request, ApplicationPipeline $pipeline)
   {
-    $pipeline->update($request->validated());
+    DB::beginTransaction();
+    try {
+      $pipeline->update($request->validated());
 
-    return $this->sendRes('Pipeline updated successfully', ['event' => 'table_reload', 'table_id' => 'application-pipelines-table', 'close' => 'globalModal']);
+      $is_default_set = false;
+      foreach ($request->stages as $i => $stage) {
+        $pipeline->stages()->updateOrCreate(['id' => @$stage['id']], ['is_default' => !$is_default_set && boolVal(@$stage['is_default']), 'order' => $i] + $stage);
+      }
+
+      // delete stages which are not in request
+      $pipeline->stages()->whereNotIn('id', array_column($request->stages, 'id'))->delete();
+      DB::commit();
+
+      return $this->sendRes('Pipeline updated successfully', ['event' => 'table_reload', 'table_id' => 'application-pipelines-table', 'close' => 'globalModal']);
+    } catch (\Exception $e) {
+      DB::rollback();
+      return $this->sendError($e->getMessage());
+    }
   }
 
   public function destroy(ApplicationPipeline $pipeline)
   {
+    if($pipeline->applications()->count() > 0) {
+      return $this->sendError('Pipeline is in use');
+    }
+    $pipeline->stages()->delete();
     $pipeline->delete();
 
     return $this->sendRes('Pipeline deleted successfully', ['event' => 'table_reload', 'table_id' => 'application-pipelines-table']);
